+++ CHANGES.rst
@@ -91,10 +91,13 @@ v1.0.13
 -----------
 * Upgrade ovp-users and ovp-projects requirements
 
-v1.0.14[unreleased]
+v1.0.14
 -----------
 * Remove 'ordered' param in favor of OrderingFilter
 * Implement project ordering by 'relevance'
 * Add result exclusion through settings
 * Add AND operator to skills and causes filter
 * Optimize UserSearch query
+
+v1.0.15[unreleased]
+-----------